requirements.txt
Django==5.0.2
pillow==10.2.0
django-autoslug==1.9.9
djangorestframework==3.14.0
django-filter==23.5
django-debug-toolbar==4.3.0
argon2-cffi==23.1.0
django-sortedm2m==3.1.1
pytils==0.4.1
django-admin-interface==0.26.0
django-crispy-forms==2.1
